+++ fixtures/autorun/some_test.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "megatest/autorun"
+
+class SomeTest < Megatest::Test
+  test "something" do
+    assert_equal 4, 1 + 1
+  end
+end